Mac ARM64 integration (#19571)
* Update Mac platform to support multi-arch detection
* New Mac arm64 specific cmake files
* Remove EditorCommon.Headers causing issues with Qt6 moving and Mac

* Fix QAbstractNativeEventFilter::nativeEventFilter Signature
* Update deprecated constructor call to QMouseEvent
* Update deprecated QHoverEvent->pos() to ->position)
* Fix issue where file that needs mocing 'QtEditorApplication_mac.h' was not in any cmake file list, thus not being auto-moc'd
* Removed unneeded include for that moc
* Added static_cast<int> in numerous places to fix error "Implicit conversion loses integer precision: 'qsizetype' (aka 'long long') to 'int'"

+++ Code/Editor/AssetImporter/UI/SelectDestinationDialog.cpp
@@ -229,7 +229,7 @@ void SelectDestinationDialog::ValidatePath()
     else
     {
         QString destinationDirectory = DestinationDirectory();
-        int strLength = destinationDirectory.length();
+        int strLength = static_cast<int>(destinationDirectory.length());
 
         // store the updated acceptable destination directory into the registry,
         // so that when users manually modify the directory,
