+++ Gems/EMotionFX/Code/Source/Editor/Plugins/SimulatedObject/SimulatedObjectSelectionWidget.cpp
@@ -106,7 +106,7 @@ namespace EMStudio
         QList<QTreeWidgetItem*> selectedItems = m_treeWidget->selectedItems();
 
         m_selectedSimulatedObjectNames.clear();
-        const uint32 numSelectedItems = selectedItems.count();
+        const uint32 numSelectedItems = static_cast<int>(selectedItems.count());
         m_selectedSimulatedObjectNames.reserve(numSelectedItems);
 
         // Iterate through the selected items in the tree widget.
